+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added user settings for default waveform colors
 - Added Epoch time (ns) number format
 - Netlist search better displays variable or scope name and instance path @DGGua
+- Multi-bit renderer at larger row heights now only vertically scales
 - Fixed status bar not updating when setting alt marker, or selecting the last value change
 - Update wellen to 0.20.3