+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Netlist search better displays variable or scope name and instance path @DGGua
 - Multi-bit renderer at larger row heights now only vertically scales
 - Fixed status bar not updating when setting alt marker, or selecting the last value change
-- Update wellen to 0.20.3
+- Update wellen to 0.20.4
 
 ## 1.5.1 - 3/9/2026 - Latest Release